+++ src/aux-common/Formulas/formula-lib.ts
@@ -569,9 +569,10 @@ function create(parent: Bot | string, ...datas: Mod[]) {
  * Combines the two given files.
  * @param first The first bot.
  * @param second The second bot.
+ * @param argument The argument to include in the script calls.
  */
-function combine(first: Bot | string, second: Bot | string) {
-    event(COMBINE_ACTION_NAME, [first, second]);
+function combine(first: Bot | string, second: Bot | string, argument?: any) {
+    event(COMBINE_ACTION_NAME, [first, second], argument);
 }
 
 /**